requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
djongo==1.3.6
Pillow==8.3.1
PyJWT==2.1.0
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
